feat: update test-harness
