+++ packages/taiko-client-rs/crates/driver/tests/dual_driver_e2e.rs
@@ -1,304 +1,44 @@
 //! E2E test verifying P2P gossip propagation between two drivers.
 
-use std::{
-    io::Write,
-    net::{IpAddr, Ipv4Addr, SocketAddr},
-    sync::Arc,
-    time::Duration,
-};
+use std::{sync::Arc, time::Duration};
 
-use alethia_reth_consensus::eip4396::{
-    SHASTA_INITIAL_BASE_FEE, calculate_next_block_eip4396_base_fee,
-};
-use alloy_consensus::{
-    EthereumTypedTransaction, SignableTransaction, TxEip1559, TxEnvelope,
-    transaction::SignerRecoverable,
-};
-use alloy_eips::{BlockId, BlockNumberOrTag, eip2718::Encodable2718};
-use alloy_primitives::{Address, B256, Bytes, TxKind, U256};
+use alloy_consensus::transaction::SignerRecoverable;
+use alloy_primitives::{Address, U256};
 use alloy_provider::Provider;
-use alloy_rlp::encode as rlp_encode;
-use alloy_rpc_types::eth::Block as RpcBlock;
-use alloy_signer::Signer;
-use alloy_signer_local::PrivateKeySigner;
 use anyhow::{Context, Result, anyhow, ensure};
 use driver::{
     DriverConfig,
-    derivation::pipeline::shasta::anchor::{AnchorTxConstructor, AnchorV4Input},
     jsonrpc::DriverRpcServer,
     sync::{SyncStage, event::EventSyncer},
 };
-use flate2::{Compression, write::ZlibEncoder};
 use preconfirmation_client::{
     DriverClient, PreconfirmationClient, PreconfirmationClientConfig,
-    codec::ZlibTxListCodec,
     driver_interface::{JsonRpcDriverClient, JsonRpcDriverClientConfig},
-    subscription::PreconfirmationEvent,
 };
-use preconfirmation_net::{InMemoryStorage, LocalValidationAdapter, P2pConfig, P2pNode};
-use preconfirmation_types::{
-    Bytes20, Bytes32, MAX_TXLIST_BYTES, PreconfCommitment, Preconfirmation, RawTxListGossip,
-    SignedCommitment, TxListBytes, address_to_bytes20, keccak256_bytes, sign_commitment,
-    u256_to_uint256, uint256_to_u256,
-};
+use preconfirmation_net::{InMemoryStorage, LocalValidationAdapter, P2pNode};
+use preconfirmation_types::uint256_to_u256;
 use rpc::{
     SubscriptionSource,
     client::{Client, ClientConfig, connect_http_with_timeout, read_jwt_secret},
 };
-use secp256k1::SecretKey;
 use serial_test::serial;
 use test_context::test_context;
 use test_harness::{
-    BeaconStubServer, PRIORITY_FEE_GWEI, ShastaEnv,
-    preconfirmation::{SafeTipDriverClient, StaticLookaheadResolver},
+    BeaconStubServer, ShastaEnv, build_anchor_tx_bytes, build_signed_transfer,
+    compute_next_block_base_fee, fetch_block_by_number,
+    preconfirmation::{
+        SafeTipDriverClient, StaticLookaheadResolver, build_publish_payloads_with_txs,
+        derive_signer, test_p2p_config, wait_for_commitment_and_txlist, wait_for_peer_connected,
+    },
+    wait_for_block,
 };
-use tokio::{
-    spawn,
-    sync::oneshot,
-    task::JoinHandle,
-    time::{Instant, sleep},
-};
+use tokio::{spawn, sync::oneshot, task::JoinHandle};
 use tracing::{info, warn};
 use url::Url;
 
-/// Creates a local-only P2P config for tests (ephemeral ports, discovery disabled).
-fn test_p2p_config() -> P2pConfig {
-    let localhost_ephemeral = SocketAddr::new(IpAddr::V4(Ipv4Addr::LOCALHOST), 0);
-    let chain_id =
-        std::env::var("L2_CHAIN_ID").ok().and_then(|v| v.parse().ok()).unwrap_or(167_001);
-
-    P2pConfig {
-        chain_id,
-        listen_addr: localhost_ephemeral,
-        discovery_listen: localhost_ephemeral,
-        enable_discovery: false,
-        ..P2pConfig::default()
-    }
-}
-
-/// Signed transfer transaction with expected hash and sender for assertions.
-struct TransferPayload {
-    raw_bytes: Bytes,
-    hash: B256,
-    from: Address,
-}
-
-/// Computes the expected base fee for the next block using EIP-4396 rules.
-async fn compute_next_block_base_fee<P>(provider: &P, parent_block_number: u64) -> Result<u64>
-where
-    P: Provider + Send + Sync,
-{
-    let parent_block = provider
-        .get_block_by_number(BlockNumberOrTag::Number(parent_block_number))
-        .await?
-        .ok_or_else(|| anyhow!("missing block {parent_block_number}"))?;
-    let parent_header = parent_block.header.inner;
-
-    if parent_header.number == 0 {
-        return Ok(SHASTA_INITIAL_BASE_FEE);
-    }
-
-    let grandparent = provider
-        .get_block_by_number(BlockNumberOrTag::Number(parent_block_number.saturating_sub(1)))
-        .await?
-        .ok_or_else(|| anyhow!("missing grandparent block"))?;
-    parent_header.base_fee_per_gas.ok_or_else(|| anyhow!("parent base fee missing"))?;
-
-    let time_delta = parent_header.timestamp.saturating_sub(grandparent.header.inner.timestamp);
-    Ok(calculate_next_block_eip4396_base_fee(&parent_header, time_delta))
-}
-
-/// Builds and signs an EIP-1559 transfer transaction.
-async fn build_signed_transfer<P>(
-    provider: &P,
-    block_number: u64,
-    private_key: &str,
-    to: Address,
-    value: U256,
-) -> Result<TransferPayload>
-where
-    P: Provider + Send + Sync,
-{
-    let signer: PrivateKeySigner = private_key.parse()?;
-    let from = signer.address();
-
-    let nonce = provider
-        .get_transaction_count(from)
-        .block_id(BlockId::Number(BlockNumberOrTag::Pending))
-        .await?;
-    let chain_id = provider.get_chain_id().await?;
-    let base_fee = compute_next_block_base_fee(provider, block_number.saturating_sub(1)).await?;
-
-    let tx = TxEip1559 {
-        chain_id,
-        nonce,
-        max_fee_per_gas: PRIORITY_FEE_GWEI + u128::from(base_fee),
-        max_priority_fee_per_gas: PRIORITY_FEE_GWEI,
-        gas_limit: 21_000,
-        to: TxKind::Call(to),
-        value,
-        ..Default::default()
-    };
-
-    let signature = signer.sign_hash(&tx.signature_hash()).await?;
-    let envelope = TxEnvelope::new_unhashed(EthereumTypedTransaction::Eip1559(tx), signature);
-
-    Ok(TransferPayload { raw_bytes: envelope.encoded_2718().into(), hash: *envelope.hash(), from })
-}
-
-/// Constructs the anchor transaction bytes for a preconfirmation block.
-async fn build_anchor_tx_bytes<P>(
-    client: &Client<P>,
-    parent_hash: B256,
-    block_number: u64,
-    base_fee: u64,
-) -> Result<Bytes>
-where
-    P: Provider + Clone + Send + Sync + 'static,
-{
-    let anchor_block_number = client.l1_provider.get_block_number().await?;
-    let anchor_block = client
-        .l1_provider
-        .get_block_by_number(BlockNumberOrTag::Number(anchor_block_number))
-        .await?
-        .ok_or_else(|| anyhow!("missing L1 anchor block {anchor_block_number}"))?;
-
-    let constructor = AnchorTxConstructor::new(client.clone()).await?;
-    let tx = constructor
-        .assemble_anchor_v4_tx(
-            parent_hash,
-            AnchorV4Input {
-                anchor_block_number,
-                anchor_block_hash: anchor_block.header.hash,
-                anchor_state_root: anchor_block.header.inner.state_root,
-                l2_height: block_number,
-                base_fee: U256::from(base_fee),
-            },
-        )
-        .await?;
-
-    Ok(tx.encoded_2718().into())
-}
-
-/// Assembles a compressed txlist and signed commitment for P2P gossip.
-fn build_publish_payloads(
-    signer_sk: &SecretKey,
-    signer: Address,
-    submission_window_end: U256,
-    block_number: U256,
-    timestamp: u64,
-    gas_limit: u64,
-    raw_tx_bytes: Vec<Bytes>,
-) -> Result<(RawTxListGossip, SignedCommitment)> {
-    let tx_list_items: Vec<Vec<u8>> = raw_tx_bytes.iter().map(|tx| tx.to_vec()).collect();
-    let tx_list = rlp_encode(&tx_list_items);
-
-    let first_byte = *tx_list.first().ok_or_else(|| anyhow!("empty tx list encoding"))?;
-    ensure!(first_byte >= 0xc0, "tx list is not an RLP list (first byte 0x{first_byte:02x})");
-
-    let mut encoder = ZlibEncoder::new(Vec::new(), Compression::default());
-    Write::write_all(&mut encoder, &tx_list)?;
-    let compressed = encoder.finish()?;
-
-    let txlist_bytes = TxListBytes::try_from(compressed)
-        .map_err(|(_, err)| anyhow!("txlist bytes error: {err}"))?;
-
-    let codec = ZlibTxListCodec::new(MAX_TXLIST_BYTES);
-    let decoded = codec.decode(txlist_bytes.as_ref()).context("decode txlist before publishing")?;
-    ensure!(decoded.len() == raw_tx_bytes.len(), "decoded txlist length mismatch");
-
-    let raw_tx_list_hash = Bytes32::try_from(keccak256_bytes(txlist_bytes.as_ref()).to_vec())
-        .map_err(|(_, err)| anyhow!("txlist hash error: {err}"))?;
-
-    let txlist =
-        RawTxListGossip { raw_tx_list_hash: raw_tx_list_hash.clone(), txlist: txlist_bytes };
-
-    let preconf = Preconfirmation {
-        eop: false,
-        block_number: u256_to_uint256(block_number),
-        timestamp: u256_to_uint256(U256::from(timestamp)),
-        gas_limit: u256_to_uint256(U256::from(gas_limit)),
-        proposal_id: u256_to_uint256(block_number),
-        coinbase: address_to_bytes20(signer),
-        submission_window_end: u256_to_uint256(submission_window_end),
-        raw_tx_list_hash,
-        ..Default::default()
-    };
-
-    let commitment = PreconfCommitment { preconf, slasher_address: Bytes20::default() };
-    let signature = sign_commitment(&commitment, signer_sk)?;
-
-    Ok((txlist, SignedCommitment { commitment, signature }))
-}
-
-/// Waits for a peer connection event.
-async fn wait_for_peer_connected(
-    events: &mut tokio::sync::broadcast::Receiver<PreconfirmationEvent>,
-) {
-    loop {
-        match events.recv().await {
-            Ok(PreconfirmationEvent::PeerConnected(_)) => return,
-            Ok(_) => continue,
-            Err(err) => panic!("preconfirmation event stream closed: {err}"),
-        }
-    }
-}
-
-/// Waits for both a commitment and its transaction list to be received.
-async fn wait_for_commitment_and_txlist(
-    events: &mut tokio::sync::broadcast::Receiver<PreconfirmationEvent>,
-) {
-    let mut saw_commitment = false;
-    let mut saw_txlist = false;
-
-    while !(saw_commitment && saw_txlist) {
-        match events.recv().await {
-            Ok(PreconfirmationEvent::NewCommitment(_)) => saw_commitment = true,
-            Ok(PreconfirmationEvent::NewTxList(_)) => saw_txlist = true,
-            Ok(_) => continue,
-            Err(err) => panic!("preconfirmation event stream closed: {err}"),
-        }
-    }
-}
-
-/// Fetches a block by number with full transaction details.
-async fn fetch_block_by_number<P>(provider: &P, block_number: u64) -> Result<RpcBlock<TxEnvelope>>
-where
-    P: Provider + Send + Sync,
-{
-    provider
-        .get_block_by_number(BlockNumberOrTag::Number(block_number))
-        .full()
-        .await?
-        .map(|b| b.map_transactions(TxEnvelope::from))
-        .ok_or_else(|| anyhow!("missing block {block_number}"))
-}
-
-/// Polls for a block until it appears or timeout expires.
-async fn wait_for_block<P>(
-    provider: &P,
-    block_number: u64,
-    timeout: Duration,
-) -> Result<RpcBlock<TxEnvelope>>
-where
-    P: Provider + Send + Sync,
-{
-    let deadline = Instant::now() + timeout;
-
-    loop {
-        if Instant::now() >= deadline {
-            return Err(anyhow!("timed out waiting for block {block_number}"));
-        }
-
-        if let Ok(Some(block)) =
-            provider.get_block_by_number(BlockNumberOrTag::Number(block_number)).full().await
-        {
-            return Ok(block.map_transactions(TxEnvelope::from));
-        }
-
-        sleep(Duration::from_millis(200)).await;
-    }
-}
+// ============================================================================
+// Driver Instance Helper
+// ============================================================================
 
 /// Running driver instance with its RPC server and background tasks.
 struct DriverInstance {
@@ -321,8 +61,8 @@ impl DriverInstance {
 
         let client_config = ClientConfig {
             l1_provider_source: l1_source.clone(),
-            l2_provider_url: l2_http.clone().into(),
-            l2_auth_provider_url: l2_auth.clone().into(),
+            l2_provider_url: l2_http.clone(),
+            l2_auth_provider_url: l2_auth.clone(),
             jwt_secret: jwt_secret_path.to_path_buf(),
             inbox_address,
         };
@@ -369,13 +109,17 @@ impl DriverInstance {
     }
 }
 
+// ============================================================================
+// Test
+// ============================================================================
+
 /// Tests that P2P gossip propagates to multiple drivers producing identical blocks.
 #[test_context(ShastaEnv)]
 #[serial]
 #[test_log::test(tokio::test(flavor = "multi_thread"))]
 async fn dual_driver_p2p_gossip_syncs_both_nodes(env: &mut ShastaEnv) -> Result<()> {
-    let l2_http_1: Url = env.l2_http_1.clone().into();
-    let l2_auth_1: Url = env.l2_auth_1.clone().into();
+    let l2_http_1: Url = env.l2_http_1.clone();
+    let l2_auth_1: Url = env.l2_auth_1.clone();
 
     let l1_http = std::env::var("L1_HTTP")?;
 
@@ -383,8 +127,8 @@ async fn dual_driver_p2p_gossip_syncs_both_nodes(env: &mut ShastaEnv) -> Result<
 
     info!("starting driver 1 (L2 node 0)");
     let driver1 = DriverInstance::start(
-        &env.l2_http.clone().into(),
-        &env.l2_auth.clone().into(),
+        &env.l2_http_0.clone(),
+        &env.l2_auth_0.clone(),
         &env.l1_source,
         &env.jwt_secret,
         env.inbox_address,
@@ -405,13 +149,13 @@ async fn dual_driver_p2p_gossip_syncs_both_nodes(env: &mut ShastaEnv) -> Result<
     .await
     .context("starting driver 2")?;
 
-    let l2_provider_1 = connect_http_with_timeout(l2_http_1.clone().into());
+    let l2_provider_1 = connect_http_with_timeout(l2_http_1.clone());
 
     let driver1_client_cfg = JsonRpcDriverClientConfig::with_http_endpoint(
         driver1.http_url().parse()?,
         env.jwt_secret.clone(),
         l1_http.parse()?,
-        env.l2_http.to_string().parse()?,
+        env.l2_http_0.to_string().parse()?,
         env.inbox_address,
     );
     let driver1_client =
@@ -427,10 +171,7 @@ async fn dual_driver_p2p_gossip_syncs_both_nodes(env: &mut ShastaEnv) -> Result<
     let driver2_client =
         SafeTipDriverClient::new(JsonRpcDriverClient::new(driver2_client_cfg).await?);
 
-    let signer_sk = SecretKey::from_slice(&[1u8; 32]).expect("secret key");
-    let signer = preconfirmation_types::public_key_to_address(
-        &secp256k1::PublicKey::from_secret_key(&secp256k1::Secp256k1::new(), &signer_sk),
-    );
+    let (signer_sk, signer) = derive_signer(1);
 
     let submission_window_end = U256::from(1000u64);
     let event_sync_tip = driver1_client.event_sync_tip().await?;
@@ -504,14 +245,14 @@ async fn dual_driver_p2p_gossip_syncs_both_nodes(env: &mut ShastaEnv) -> Result<
     .await?;
 
     let test_key = std::env::var("TEST_ACCOUNT_PRIVATE_KEY")?;
-    let test_signer: PrivateKeySigner = test_key.parse()?;
+    let test_signer: alloy_signer_local::PrivateKeySigner = test_key.parse()?;
     let test_address = test_signer.address();
     let test_balance = env.client.l2_provider.get_balance(test_address).await?;
 
     let mut transfers = Vec::new();
     if test_balance.is_zero() {
         let funder_key = std::env::var("PRIVATE_KEY")?;
-        let funder_signer: PrivateKeySigner = funder_key.parse()?;
+        let funder_signer: alloy_signer_local::PrivateKeySigner = funder_key.parse()?;
         let funder_balance = env.client.l2_provider.get_balance(funder_signer.address()).await?;
         let fund_amount = U256::from(1_000_000_000_000_000_000u128);
         ensure!(funder_balance > fund_amount, "funder balance too low to seed test account");
@@ -542,7 +283,7 @@ async fn dual_driver_p2p_gossip_syncs_both_nodes(env: &mut ShastaEnv) -> Result<
     let mut txlist_transactions = vec![anchor_tx];
     txlist_transactions.extend(transfers.iter().map(|t| t.raw_bytes.clone()));
 
-    let (txlist, signed_commitment) = build_publish_payloads(
+    let (txlist, signed_commitment) = build_publish_payloads_with_txs(
         &signer_sk,
         signer,
         submission_window_end,
