+++ packages/taiko-client-rs/crates/driver/tests/proposer_driver_e2e.rs
@@ -27,8 +27,8 @@ use tracing::{info, warn};
 fn client_config(env: &ShastaEnv) -> ClientConfig {
     ClientConfig {
         l1_provider_source: env.l1_source.clone(),
-        l2_provider_url: env.l2_http.clone(),
-        l2_auth_provider_url: env.l2_auth.clone(),
+        l2_provider_url: env.l2_http_0.clone(),
+        l2_auth_provider_url: env.l2_auth_0.clone(),
         jwt_secret: env.jwt_secret.clone(),
         inbox_address: env.inbox_address,
     }
