+++ packages/taiko-client-rs/crates/driver/tests/shasta_event_sync.rs
@@ -25,8 +25,8 @@ async fn syncs_shasta_proposal_into_l2(env: &mut ShastaEnv) -> Result<()> {
     let proposer_client = Client::new_with_wallet(
         ClientConfig {
             l1_provider_source: env.l1_source.clone(),
-            l2_provider_url: env.l2_http.clone(),
-            l2_auth_provider_url: env.l2_auth.clone(),
+            l2_provider_url: env.l2_http_0.clone(),
+            l2_auth_provider_url: env.l2_auth_0.clone(),
             jwt_secret: env.jwt_secret.clone(),
             inbox_address: env.inbox_address,
         },
@@ -72,8 +72,8 @@ async fn syncs_shasta_proposal_into_l2(env: &mut ShastaEnv) -> Result<()> {
     let driver_config = DriverConfig::new(
         ClientConfig {
             l1_provider_source: env.l1_source.clone(),
-            l2_provider_url: env.l2_http.clone(),
-            l2_auth_provider_url: env.l2_auth.clone(),
+            l2_provider_url: env.l2_http_0.clone(),
+            l2_auth_provider_url: env.l2_auth_0.clone(),
             jwt_secret: env.jwt_secret.clone(),
             inbox_address: env.inbox_address,
         },
